auth/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.7
PyJWT==2.8.0
bcrypt==4.1.2
redis==5.0.1
python-dotenv==1.0.0
